source/retail_ai_assistant_app/requirements.txt
streamlit==1.37.0
boto3
streamlit-cognito-auth==1.3.1
python-dotenv
pandas
requests